Add support for non-USD denominated stocks
Previously if the user were to add a stock priced in e.g. GBP, the
stock's value would simply be saved as if it were in USD.

I have added checking stages to stock_check.py which verify that
the stock is priced either in USD or a currency the exchange rate
for which is stored in models.py. If the exchange rate is found
the stock's value is converted to USD.