Add USD to JS function; fix non-USD holding bug
Upon testing I realised that when the repo is downloaded as is and
run locally, there is no USD Rate object by default. Therefore the
selectCurrency() function was not working properly. I have added
some hardcoding to the function to fix this. In this case I think
hardcoding is acceptable as 1 USD will always = 1 USD, and it's more
user friendly than asking users to manually add a USD object in the
django shell before running the program!

I also found that GBP-denominated stocks were still being added with
their GBP value as USD. I have made some changes to stock_check.py
to address this.