Change Rate / User to many-to-many relationship
Prior to this change, the landing page simply displayed all stored
Rate objects, so all Users saw all Rates that had been added,
including by other users.

I have added the Userrates model to act as a junction table. When
a user adds a rate, the RateCreateView first checks whether a
ticker is stored in the database, and creates the Rate object if
not.

A new Userrates object is then created linking the User to the Rate.
On the landing page the User now only sees Rates that have been
linked to themselves through this junction table.

In addition, HoldingUpdateAllView now only updates Rates that are
linked to the User making the request.

This allows the app to scale much more effeciently to a large number
of Users.